Moved more common stuff to mixins
