+++ esteid/authentication/views.py
@@ -1,13 +1,10 @@
 import logging
-from http import HTTPStatus
 from typing import Type, TYPE_CHECKING
 
-from django.core.exceptions import ValidationError as DjangoValidationError
-from django.http import Http404, HttpRequest, JsonResponse
-from django.utils.translation import gettext
+from django.http import HttpRequest, JsonResponse
 
-from esteid.exceptions import ActionInProgress, CanceledByUser, EsteidError, InvalidParameters
-from esteid.mixins import DjangoRestCompatibilityMixin
+from esteid.exceptions import ActionInProgress
+from esteid.mixins import DjangoRestCompatibilityMixin, SessionViewMixin
 
 from .authenticator import Authenticator
 from .types import AuthenticationResult
@@ -33,12 +30,7 @@ if TYPE_CHECKING:
 logger = logging.getLogger(__name__)
 
 
-class AuthenticationViewMixin:
-    class Status:
-        ERROR = "error"
-        PENDING = "pending"
-        SUCCESS = "success"
-
+class AuthenticationViewMixin(SessionViewMixin):
     # this comes from the `url()` definition as in `View.as_view(authentication_method='...')`
     authentication_method: str = None
 
@@ -104,55 +96,6 @@ class AuthenticationViewMixin:
 
         return JsonResponse({**result, "status": self.Status.SUCCESS})
 
-    def report_error(self, e: EsteidError):
-        return JsonResponse({"status": self.Status.ERROR, **e.get_user_error()}, status=e.status)
-
-    def handle_user_cancel(self):
-        pass
-
-    def handle_errors(self, e: Exception, stage="start"):
-        if isinstance(e, EsteidError):
-            if isinstance(e, CanceledByUser):
-                self.handle_user_cancel()
-
-            # Do not log user input related errors
-            if not isinstance(e, InvalidParameters):
-                logger.exception("Failed to %s signing session.", stage)
-            return self.report_error(e)
-
-        if isinstance(e, (Http404, DRFValidationError)):
-            raise e
-
-        if isinstance(e, DjangoValidationError):
-            return JsonResponse(
-                {"status": self.Status.ERROR, "error": e.__class__.__name__, "message": str(e)},
-                status=HTTPStatus.CONFLICT,
-            )
-
-        logger.exception("Failed to %s signing session.", stage)
-        return JsonResponse(
-            {"status": self.Status.ERROR, "error": "Internal error", "message": gettext("Internal server error")},
-            status=HTTPStatus.INTERNAL_SERVER_ERROR,
-        )
-
-    def post(self, request, *args, **kwargs):
-        """
-        Handles session start requests
-        """
-        try:
-            return self.start_session(request, *args, **kwargs)
-        except Exception as e:
-            return self.handle_errors(e, stage="start")
-
-    def patch(self, request, *args, **kwargs):
-        """
-        Handles session finish requests
-        """
-        try:
-            return self.finish_session(request, *args, **kwargs)
-        except Exception as e:
-            return self.handle_errors(e, stage="finish")
-
 
 class AuthenticationViewRestMixin(AuthenticationViewMixin):
     """
@@ -166,17 +109,3 @@ class AuthenticationViewDjangoMixin(DjangoRestCompatibilityMixin, Authentication
 
     Adds `data` attribute to the request with the POST or JSON data.
     """
-
-    def post(self, request: "RequestType", *args, **kwargs):
-        """
-        Handles session start requests
-        """
-        request.data = self.parse_request(request)
-        return super().start_session(request, *args, **kwargs)
-
-    def patch(self, request: "RequestType", *args, **kwargs):
-        """
-        Handles session finish requests
-        """
-        request.data = self.parse_request(request)
-        return super().finish_session(request, *args, **kwargs)
